refactor: move alpha and datasource to separate modules, add `main.py`
